# apps/web/openapi/mediamop-openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyUpdateIn": {
        "additionalProperties": false,
        "description": "Body for POST /suite/apply-update.",
        "properties": {
          "csrf_token": {
            "minLength": 1,
            "title": "Csrf Token",
            "type": "string"
          }
        },
        "required": [
          "csrf_token"
        ],
        "title": "ApplyUpdateIn",
        "type": "object"
      },
...
      "ConfigurationBundleImportIn": {
        "additionalProperties": false,
        "description": "Restore suite + module settings from a prior configuration bundle export.",
        "properties": {
          "bundle": {
            "additionalProperties": true,
            "title": "Bundle",
            "type": "object"
          },
          "csrf_token": {
            "minLength": 1,
            "title": "Csrf Token",
            "type": "string"
          }
        },
        "required": [
          "csrf_token",
          "bundle"
        ],
        "title": "ConfigurationBundleImportIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuiteLogCountsOut": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "title": "Error",
            "type": "integer"
          },
          "information": {
            "title": "Information",
            "type": "integer"
          },
          "warning": {
            "title": "Warning",
            "type": "integer"
          }
        },
        "required": [
          "error",
          "warning",
          "information"
        ],
        "title": "SuiteLogCountsOut",
        "type": "object"
      },
...
      "SuiteLogsOut": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/SuiteLogCountsOut"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SuiteLogEntryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "counts"
        ],
        "title": "SuiteLogsOut",
        "type": "object"
      },
...
      "SuiteOperationalHistoryResetIn": {
        "additionalProperties": false,
        "description": "Body for ``POST /suite/operational-history/reset``.",
        "properties": {
          "confirm": {
            "maxLength": 32,
            "minLength": 5,
            "title": "Confirm",
            "type": "string"
          },
          "csrf_token": {
            "minLength": 1,
            "title": "Csrf Token",
            "type": "string"
          }
        },
        "required": [
          "csrf_token",
          "confirm"
        ],
        "title": "SuiteOperationalHistoryResetIn",
        "type": "object"
      },
      "SuiteOperationalHistoryResetOut": {
        "additionalProperties": false,
        "description": "Counts removed by an explicit operator reset.",
        "properties": {
          "activity_events_deleted": {
            "minimum": 0.0,
            "title": "Activity Events Deleted",
            "type": "integer"
          },
          "pruner_jobs_deleted": {
            "minimum": 0.0,
            "title": "Pruner Jobs Deleted",
            "type": "integer"
          },
          "refiner_jobs_deleted": {
            "minimum": 0.0,
            "title": "Refiner Jobs Deleted",
            "type": "integer"
          },
          "status": {
            "minLength": 1,
            "title": "Status",
            "type": "string"
          },
          "total_deleted": {
            "minimum": 0.0,
            "title": "Total Deleted",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "activity_events_deleted",
          "refiner_jobs_deleted",
          "pruner_jobs_deleted",
          "total_deleted"
        ],
        "title": "SuiteOperationalHistoryResetOut",
        "type": "object"
      },
...
      "SuiteSettingsOut": {
        "additionalProperties": false,
        "description": "Values stored in ``suite_settings`` (editable in the app).",
        "properties": {
          "app_timezone": {
            "description": "Suite-wide timezone label used for date/time displays that follow app timezone.",
            "maxLength": 120,
            "title": "App Timezone",
            "type": "string"
          },
          "configuration_backup_enabled": {
            "description": "Whether server-side automatic configuration snapshots are enabled.",
            "title": "Configuration Backup Enabled",
            "type": "boolean"
          },
          "configuration_backup_interval_hours": {
            "description": "Minimum hours between automatic configuration snapshots.",
            "maximum": 720.0,
            "minimum": 1.0,
            "title": "Configuration Backup Interval Hours",
            "type": "integer"
          },
          "configuration_backup_last_run_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UTC timestamp of the last successful automatic configuration snapshot run.",
            "title": "Configuration Backup Last Run At"
          },
          "configuration_backup_preferred_time": {
            "description": "Preferred local backup time in HH:MM for daily-style automatic snapshots.",
            "maxLength": 5,
            "minLength": 5,
            "title": "Configuration Backup Preferred Time",
            "type": "string"
          },
          "log_retention_days": {
            "description": "How long persisted system logs are kept before automatic cleanup.",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Log Retention Days",
            "type": "integer"
          },
          "product_display_name": {
            "description": "Shown in the sidebar and settings.",
            "maxLength": 120,
            "title": "Product Display Name",
            "type": "string"
          },
          "setup_wizard_state": {
            "description": "First-run wizard state: pending, skipped, or completed.",
            "maxLength": 32,
            "minLength": 1,
            "title": "Setup Wizard State",
            "type": "string"
          },
          "signed_in_home_notice": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional short message on the home dashboard for signed-in users.",
            "title": "Signed In Home Notice"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "product_display_name",
          "setup_wizard_state",
          "app_timezone",
          "log_retention_days",
          "configuration_backup_enabled",
          "configuration_backup_interval_hours",
          "configuration_backup_preferred_time",
          "updated_at"
        ],
        "title": "SuiteSettingsOut",
        "type": "object"
      },
...
      "UpdateStateOut": {
        "additionalProperties": false,
        "description": "Current state of any downloaded-but-not-yet-applied update (read from update-state.json).",
        "properties": {
          "downloaded": {
            "default": false,
            "title": "Downloaded",
            "type": "boolean"
          },
          "pending_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Version"
          }
        },
        "title": "UpdateStateOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MediaMop API",
    "version": "2.4.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/suite/apply-update": {
      "post": {
        "description": "Signal the tray to restart and apply the pending update.",
        "operationId": "post_suite_apply_update_api_v1_suite_apply_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateStateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Suite Apply Update",
        "tags": [
          "suite"
        ]
      }
    },
...
    "/api/v1/suite/configuration-backups/{backup_id}/download": {
      "get": {
        "operationId": "download_configuration_backup_api_v1_suite_configuration_backups__backup_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "backup_id",
            "required": true,
            "schema": {
              "title": "Backup Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Configuration Backup",
        "tags": [
          "suite"
        ]
      }
    },
    "/api/v1/suite/configuration-bundle": {
      "get": {
        "description": "Export suite + module configuration as JSON (operators/admins only \u2014 contains secrets).",
        "operationId": "get_configuration_bundle_api_v1_suite_configuration_bundle_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Configuration Bundle Api V1 Suite Configuration Bundle Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Configuration Bundle",
        "tags": [
          "suite"
        ]
      },
      "put": {
        "description": "Replace suite + module configuration from a prior export (operators/admins only).",
        "operationId": "put_configuration_bundle_api_v1_suite_configuration_bundle_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationBundleImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Configuration Bundle Api V1 Suite Configuration Bundle Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Configuration Bundle",
        "tags": [
          "suite"
        ]
      }
    },
    "/api/v1/suite/logs": {
      "get": {
        "operationId": "get_suite_logs_api_v1_suite_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "has_exception",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Exception"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteLogsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Suite Logs",
        "tags": [
          "suite"
        ]
      }
    },
...
    "/api/v1/suite/operational-history/reset": {
      "post": {
        "description": "Explicitly clear dashboard/activity history without touching settings or active work.",
        "operationId": "post_suite_operational_history_reset_api_v1_suite_operational_history_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteOperationalHistoryResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteOperationalHistoryResetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Suite Operational History Reset",
        "tags": [
          "suite"
        ]
      }
    },
...
    "/api/v1/suite/settings": {
      "get": {
        "description": "Names and notices stored in the app database (any signed-in user may read).",
        "operationId": "get_suite_settings_api_v1_suite_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Suite Settings",
        "tags": [
          "suite"
        ]
      },
      "put": {
        "description": "Update suite display text \u2014 operators and admins only.",
        "operationId": "put_suite_settings_api_v1_suite_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuiteSettingsPutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Suite Settings",
        "tags": [
          "suite"
        ]
      }
    },
    "/api/v1/suite/settings/configuration-bundle": {
      "get": {
        "description": "Export suite + module configuration as JSON (operators/admins only \u2014 contains secrets).",
        "operationId": "get_configuration_bundle_api_v1_suite_settings_configuration_bundle_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Configuration Bundle Api V1 Suite Settings Configuration Bundle Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Configuration Bundle",
        "tags": [
          "suite"
        ]
      },
      "put": {
        "description": "Replace suite + module configuration from a prior export (operators/admins only).",
        "operationId": "put_configuration_bundle_api_v1_suite_settings_configuration_bundle_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationBundleImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Configuration Bundle Api V1 Suite Settings Configuration Bundle Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Configuration Bundle",
        "tags": [
          "suite"
        ]
      }
    },
...
    "/api/v1/suite/update-settings": {
      "get": {
        "description": "Read the tray update-mode preferences from update-settings.json.",
        "operationId": "get_suite_update_settings_api_v1_suite_update_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Suite Update Settings",
        "tags": [
          "suite"
        ]
      },
      "put": {
        "description": "Persist tray update-mode preferences to update-settings.json.",
        "operationId": "put_suite_update_settings_api_v1_suite_update_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsPutIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Suite Update Settings",
        "tags": [
          "suite"
        ]
      }
    },
...
  }
}